# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BeatSight Backend",
    "description": "\n## BeatSight Web API\n\nBeatSight is an AI-powered drum beatmap generation platform. This API provides:\n\n### Features\n- **Authentication**: JWT-based auth with access/refresh tokens\n- **Songs**: Upload, manage, and organize your music library\n- **AI Jobs**: Queue songs for AI beatmap generation with real-time progress\n- **Cloud Sync**: Sync preferences and progress across devices\n- **Storage**: Secure file upload with presigned URLs\n\n### Authentication\nMost endpoints require authentication. Include the JWT token in the Authorization header:\n```\nAuthorization: Bearer <access_token>\n```\n\n### Rate Limits\n- Standard users: 100 requests/minute\n- Premium users: 500 requests/minute\n- AI job quota: Based on subscription tier\n    ",
    "contact": {
      "name": "BeatSight Support",
      "url": "https://beatsight.app/support",
      "email": "support@beatsight.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sync/preferences": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Get Preferences",
        "description": "Get current user preferences.",
        "operationId": "get_preferences_api_sync_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "sync"
        ],
        "summary": "Update Preferences",
        "description": "Update user preferences with optional optimistic locking.",
        "operationId": "update_preferences_api_sync_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sync/clients": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "List Clients",
        "description": "List all registered sync clients for the current user.",
        "operationId": "list_clients_api_sync_clients_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SyncClientResponse"
                  },
                  "type": "array",
                  "title": "Response List Clients Api Sync Clients Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Register Client",
        "description": "Register a new sync client (device).",
        "operationId": "register_client_api_sync_clients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncClientResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sync/clients/{client_id}": {
      "delete": {
        "tags": [
          "sync"
        ],
        "summary": "Remove Client",
        "description": "Remove a sync client.",
        "operationId": "remove_client_api_sync_clients__client_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/manifest": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Compare Manifest",
        "description": "Compare client manifest with cloud to determine sync actions.\n\nThis is Phase 1 of the sync protocol - discovery phase.",
        "operationId": "compare_manifest_api_sync_manifest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncManifestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManifestCompareResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/sync/manifest/{map_id}": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Update Manifest Entry",
        "description": "Update manifest entry after successful upload/download.",
        "operationId": "update_manifest_entry_api_sync_manifest__map_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Map Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Version"
            }
          },
          {
            "name": "checksum",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Checksum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Manifest Entry Api Sync Manifest  Map Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/conflicts": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "List Conflicts",
        "description": "List sync conflicts for the current user.",
        "operationId": "list_conflicts_api_sync_conflicts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_resolved",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Resolved"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConflictResponse"
                  },
                  "title": "Response List Conflicts Api Sync Conflicts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/conflicts/{conflict_id}/resolve": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Resolve Conflict",
        "description": "Resolve a sync conflict.",
        "operationId": "resolve_conflict_api_sync_conflicts__conflict_id__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conflict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Conflict Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveConflictRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/history": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Get Sync History",
        "description": "Get recent sync history.",
        "operationId": "get_sync_history_api_sync_history_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SyncLogResponse"
                  },
                  "title": "Response Get Sync History Api Sync History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/status": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Get Sync Status",
        "description": "Get overall sync status for the current user.",
        "operationId": "get_sync_status_api_sync_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferencesResponse": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "checksum": {
            "type": "string",
            "title": "Checksum"
          },
          "scroll_speed": {
            "type": "number",
            "title": "Scroll Speed"
          },
          "note_skin": {
            "type": "string",
            "title": "Note Skin"
          },
          "audio_offset_ms": {
            "type": "integer",
            "title": "Audio Offset Ms"
          },
          "visual_offset_ms": {
            "type": "integer",
            "title": "Visual Offset Ms"
          },
          "background_dim": {
            "type": "number",
            "title": "Background Dim"
          },
          "master_volume": {
            "type": "number",
            "title": "Master Volume"
          },
          "music_volume": {
            "type": "number",
            "title": "Music Volume"
          },
          "effects_volume": {
            "type": "number",
            "title": "Effects Volume"
          },
          "hitsound_volume": {
            "type": "number",
            "title": "Hitsound Volume"
          },
          "theme": {
            "type": "string",
            "title": "Theme"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "custom_settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Custom Settings"
          },
          "last_modified": {
            "type": "string",
            "format": "date-time",
            "title": "Last Modified"
          }
        },
        "type": "object",
        "required": [
          "version",
          "checksum",
          "scroll_speed",
          "note_skin",
          "audio_offset_ms",
          "visual_offset_ms",
          "background_dim",
          "master_volume",
          "music_volume",
          "effects_volume",
          "hitsound_volume",
          "theme",
          "language",
          "custom_settings",
          "last_modified"
        ],
        "title": "PreferencesResponse",
        "description": "User preferences response."
      },
      "PreferencesUpdate": {
        "properties": {
          "scroll_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scroll Speed"
          },
          "note_skin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Skin"
          },
          "audio_offset_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Offset Ms"
          },
          "visual_offset_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Offset Ms"
          },
          "background_dim": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Background Dim"
          },
          "master_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Master Volume"
          },
          "music_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Music Volume"
          },
          "effects_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effects Volume"
          },
          "hitsound_volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hitsound Volume"
          },
          "theme": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "custom_settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Settings"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version",
            "description": "For optimistic locking"
          }
        },
        "type": "object",
        "title": "PreferencesUpdate",
        "description": "Update user preferences."
      },
...
      "ResolveConflictRequest": {
        "properties": {
          "resolution": {
            "$ref": "#/components/schemas/ConflictResolution"
          }
        },
        "type": "object",
        "required": [
          "resolution"
        ],
        "title": "ResolveConflictRequest",
        "description": "Request to resolve a sync conflict."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}